Restructure module
