+++ data_safe_haven/commands/cli.py
@@ -4,8 +4,7 @@ from typing import Annotated, Optional
 
 import typer
 
-from data_safe_haven import __version__
-from data_safe_haven.console import pretty_print
+from data_safe_haven import __version__, console
 from data_safe_haven.logging import set_console_level, show_console_level
 
 from .config import config_command_group
@@ -60,7 +59,7 @@ def callback(
         show_console_level()
 
     if version:
-        pretty_print(f"Data Safe Haven {__version__}")
+        console.print(f"Data Safe Haven {__version__}")
         raise typer.Exit()
