+++ data_safe_haven/console/__init__.py
@@ -1,4 +1,9 @@
-from . import fmt, prompts
-from .pretty import pretty_print
+from .format import tabulate
+from .pretty import pretty_print as print  # noqa: A001
+from .prompts import confirm
 
-__all__ = ["fmt", "pretty_print", "prompts"]
+__all__ = [
+    "confirm",
+    "print",
+    "tabulate",
+]
