+++ data_safe_haven/console/format.py
